pyproject.toml
[tool.poetry]
name = "in-house"
version = "0.1.0"
description = ""
authors = ["Yashwanth Rama Krishna Reddy <yashwanth10005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.2"
openai = "0.28"
djangorestframework-simplejwt = "^5.3.1"
python-dotenv = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
black = "^24.4.2"
flake8 = "^7.1.0"
mypy = "^1.10.1"
pre-commit = "^3.4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
include = '\.pyi?$'
exclude = '''
/(
    \.git
  | \.venv
  | \.env
  | _build
  | build
  | dist
)/
'''

[tool.flake8]
max-line-length = 88
exclude = ".git,__pycache__,build,dist,.venv"
ignore = "E203,W503"

[tool.mypy]
python_version = 3.12
warn_return_any = true
warn_unused_configs = true

[tool.ruff]
line-length = 88
select = ["E", "F", "C", "W"]
